pyproject.toml
[project]
name = "food-recsys"
version = "0.1.0"
description = "A Food Recommendation System that suggests the best dishes based on user preferences and dietary restrictions"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.0",
    "neo4j>=5.25.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "transformers>=4.46.0",
    "torch>=2.5.0",
    "ollama>=0.6.1",
    "accelerate>=1.12.0",
]

[project.scripts]
food-recsys = "main:main"
